vhu_rpmb: implement GET_WRITE_COUNTER
Implementing the logic was fairly simple but the challenge was in
calculating the Hmac and exactly how to pass the mac value around and
use it to calculate the Mac on the final message. There are also
warnings about the alignment of a bunch of the fields in the
structure.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>